Added DAC Functionality

+++ Final_project_ESE5013/Includes/tsi.h
@@ -8,7 +8,7 @@
 #ifndef INCLUDES_TSI_H_
 #define INCLUDES_TSI_H_
 uint8_t f;
-volatile uint32_t d;
+volatile uint16_t d;
 void tsi_init(void);
 void TSI0_IRQHandler(void);
 void tsi_led(void);
